+++ wasatch/CSVLoader.py
@@ -12,6 +12,11 @@ log = logging.getLogger(__name__)
 ##
 # A file parser to deserialize one Measurement from a column-ordered CSV file.
 #
+# It looks like the basic usage is to instantiate it with a CSV pathname, call
+# load_data() with or without scalar_metadata, then access the .processed_reading
+# and .metadata elements. So, arguably this could have been a separate 
+# constructor for ProcessedReading?
+#
 # Given the similarity between the columnar CSV and "export" file formats, it 
 # would be SO TEMPTING to imagine you could easily generalize them.  I thought
 # they're just different enough that it would be a nightmare, so here we are.
@@ -21,10 +26,12 @@ log = logging.getLogger(__name__)
 # row.  In that case, wavelength and wavenumber are used directly from the input
 # data, as no wavecal coefficients or excitation are available.
 #
-# Currently this class is used by two callers:
+# Currently this class is used by at least three callers:
 #
-# - enlighten.parser.ColumFileParser
+# - enlighten.parser.ColumnFileParser
 # - wasatch.MockUSBDevice
+# - an OEM plugin
+# - maybe a shell script or two? (check enlighten/scripts)
 #
 class CSVLoader:
 
@@ -37,23 +44,24 @@ class CSVLoader:
 
         # temporarily store these if no wavecal is provided
         self.metadata = {
-            "pixel": [],
-            "wavelength": [],
-            "wavenumber": []
+            "pixel": []
         }
 
         self.headers = []
         self.processed_reading = ProcessedReading()
         self.processed_reading.reading = Reading(device_id = "LOAD:" + pathname)
 
-    def parse_metadata(self, line, scalar_metadata=False):
+    def __repr__(self):
+        return f"CSVLoader<{self.pathname}> with ProcessedReading {self.processed_reading} and metadata {self.metadata.keys()}"
+
+    def _parse_metadata(self, line, scalar_metadata=False):
         """
         MZ: I'm not sure who is using this method and wants the metadata 
         values to be returned as lists, but it unnecessarily complicates 
         ColumnFileParser, so adding the scalar option.
         """
         line = list(line)
-        key   = line[0]
+        key = line[0]
         if scalar_metadata:
             value = line[1]
         else:
@@ -61,16 +69,25 @@ class CSVLoader:
                              # Useful for cases like Declared Match,,,,,,,,
         self.metadata[key] = value
 
-    def parse_header(self, line):
+    def _parse_header(self, line):
         self.headers = [ x.lower().strip() for x in line ] # force lowercase
         if "processed" in self.headers: self.processed_reading.processed = []
         if "raw"       in self.headers: self.processed_reading.raw       = []
         if "dark"      in self.headers: self.processed_reading.dark      = []
         if "reference" in self.headers: self.processed_reading.reference = []
         if "corrected" in self.headers: self.processed_reading.processed = []
-        log.debug("parse_header: headers = %s", self.headers)
+        # log.debug("_parse_header: headers = %s", self.headers)
 
     def load_data(self, scalar_metadata=False):
+
+        # create empty arrays for all of these so we have someplace to put data as we load lines
+        self.processed_reading.processed = []
+        self.processed_reading.raw = []
+        self.processed_reading.dark = []
+        self.processed_reading.reference = []
+        self.processed_reading.wavelengths = []
+        self.processed_reading.wavenumbers = []
+
         state = "reading_metadata"
         data_rows_read = 0
         with open(self.pathname, "r", encoding=self.encoding) as infile:
@@ -81,25 +98,26 @@ class CSVLoader:
                     continue
 
                 line[-1] = line[-1].strip() # remove the \n
-                log.debug("load_data[%s]: %s", state, line)
+                # log.debug("load_data[%s]: %s", state, line)
 
                 if state == "reading_metadata":
                     
-                    # check for end of metadata (note trailing comma!)
-                    cleanup_line = lambda x: x.strip().lower()
-                    clean_line = [cleanup_line(part) for part in line]
-                    check_present = lambda x: x in clean_line
-                    contains_header = [check_present(header) 
-                                       for header in ["pixel", "wavelength", "wavenumber", "processed"]]
-                    if any(contains_header):
-                        self.parse_header(line)
+                    # check for end of metadata
+                    looks_like_header = False
+                    for tok in [ part.strip().lower() for part in line ]:
+                        if tok in ["pixel", "wavelength", "wavenumber", "processed", "intensity"]:
+                            looks_like_header = True
+                            break
+
+                    if looks_like_header:
+                        self._parse_header(line)
                         state = "reading_data"
                     else:
                         # still in metadata
-                        self.parse_metadata(line, scalar_metadata)
+                        self._parse_metadata(line, scalar_metadata)
 
                 elif state == "reading_metadata_final":
-                    self.parse_metadata(line, scalar_metadata)
+                    self._parse_metadata(line, scalar_metadata)
                 
                 elif state == "reading_data":
                     values = [x.strip() for x in line]
@@ -108,7 +126,7 @@ class CSVLoader:
                     # do not transition back
                     if not re.match(r'^[-+]?\d', values[0]):
                         state = "reading_metadata_final"
-                        self.parse_metadata(line, scalar_metadata)
+                        self._parse_metadata(line, scalar_metadata)
                         continue
 
                     # Assume each value read aligns with a known headers, but recognize that there
@@ -117,9 +135,11 @@ class CSVLoader:
                     # comma-delimited fields actually read (or the number of headers, if more values
                     # were read than had headers).
                     count = min(len(self.headers), len(values))
+                    # log.debug(f"parsing {count} fields")
 
                     for i in range(count):
                         header = self.headers[i]
+                        # log.debug(f"parsing header {i} ({header})")
 
                         # SKIP nulls.  Note we're APPENDING data to each list, so this means that 
                         # if there are blanks (rather than '0' zeros) in the MIDDLE of a column, 
@@ -127,30 +147,39 @@ class CSVLoader:
                         # improperly-associated "rows".
                         value = values[i]
                         if len(value) == 0:
+                            # log.debug(f"skipping value {i} ({value})")
                             continue
 
                         # MZ: honestly not sure if we should skip these or treat as zero
                         if value == "NA":
+                            # log.debug(f"treating value {i} ({value}) as zero")
                             value = 0
 
                         # add to array
                         array = None
-                        if   header == "processed":  array = self.processed_reading.processed
+                        if   header in ["processed", "intensity"]:  array = self.processed_reading.processed
                         elif header == "corrected":  array = self.processed_reading.processed 
                         elif header == "raw":        array = self.processed_reading.raw
                         elif header == "dark":       array = self.processed_reading.dark
                         elif header == "reference":  array = self.processed_reading.reference
-                        elif header == "pixel":      array = self.metadata['pixel']      
-                        elif header == "wavelength": array = self.metadata['wavelength'] 
-                        elif header == "wavenumber": array = self.metadata['wavenumber'] 
+                        elif header == "wavelength": array = self.processed_reading.wavelengths
+                        elif header == "wavenumber": array = self.processed_reading.wavenumbers
+                        elif header == "pixel":      array = self.metadata["pixel"]
 
                         if array is not None:
-                            # log.debug("appending to %s: %s", header, value)
+                            # log.debug(f"appending to {header}: {value}")
                             array.append(float(value))
                         else:
-                            log.error("load_data: null array?")
+                            log.error(f"load_data: null array? headers {self.headers}, header {header}, value {value}, line {line} ({self.pathname})")
 
                         if data_rows_read == 0 and header == "pixel" and int(value) != 0:
-                            self.processed_reading.first_pixel = int(value)
+                            value = int(value)
+                            # log.debug(f"first pixel = {value}")
+                            self.processed_reading.first_pixel = value
 
                     data_rows_read += 1
+
+        # clear any arrays we ended up not filling
+        self.processed_reading.post_load_cleanup()
+
+        return self.processed_reading, self.metadata